🔄 synced local 'skyvern-frontend/src/' with remote 'skyvern-frontend/src/'
- remove copy code button from code dropdown
- move CopyButton to it's own component file

https://linear.app/skyvern/issue/SKY-6041/remove-copy-script-name-icon-because-it-occupies-precious-space
<!-- ELLIPSIS_HIDDEN -->

----

> [!IMPORTANT]
> Remove copy button from code dropdown and refactor `CopyButton` into its own file.
>
>   - **Behavior**:
>     - Remove copy button from code dropdown in `WorkflowHeader.tsx`.
>   - **Refactoring**:
>     - Move `CopyButton` component to `CopyButton.tsx`.
>     - Remove `CopyButton` logic from `WorkflowHeader.tsx`.
>
> <sup>This description was created by </sup>[<img alt="Ellipsis" src="https://img.shields.io/badge/Ellipsis-blue?color=175173">](https://www.ellipsis.dev?ref=Skyvern-AI%2Fskyvern-cloud&utm_source=github&utm_medium=referral)<sup> for 9c6aeaa8d75001e7fc7bb75b13ce943aaf6efd1e. You can [customize](https://app.ellipsis.dev/Skyvern-AI/settings/summaries) this summary. It will automatically update as commits are pushed.</sup>

<!-- ELLIPSIS_HIDDEN -->
